refactor:#16 rename like domain to postlike for better clarity
